Adjust precision. Implement binary mode

+++ src/GPUGeometry.cpp
@@ -10,28 +10,28 @@ GPUGeometry::GPUGeometry(Geometry &geom)
 
 	// Allocate the memory
 
-	checkCudaErrors(cudaMalloc((void **)&normal, arraySize * sizeof(double3)));
+	checkCudaErrors(cudaMalloc((void **)&normal, arraySize * sizeof(float3)));
 
-	checkCudaErrors(cudaMalloc((void **)&vertexA, arraySize * sizeof(double3)));
+	checkCudaErrors(cudaMalloc((void **)&vertexA, arraySize * sizeof(float3)));
 
-	checkCudaErrors(cudaMalloc((void **)&edgeBA, arraySize * sizeof(double3)));
+	checkCudaErrors(cudaMalloc((void **)&edgeBA, arraySize * sizeof(float3)));
 
-	checkCudaErrors(cudaMalloc((void **)&edgeCA, arraySize * sizeof(double3)));
+	checkCudaErrors(cudaMalloc((void **)&edgeCA, arraySize * sizeof(float3)));
 
-	checkCudaErrors(cudaMalloc((void **)&center, arraySize * sizeof(double3)));
+	checkCudaErrors(cudaMalloc((void **)&center, arraySize * sizeof(float3)));
 
 	checkCudaErrors(cudaMalloc((void **)&area, arraySize * sizeof(double)));
 
 	// Copy the memory
-	checkCudaErrors(cudaMemcpy(normal, geom.normal, arraySize * sizeof(double3), cudaMemcpyHostToDevice));
+	checkCudaErrors(cudaMemcpy(normal, geom.normal, arraySize * sizeof(float3), cudaMemcpyHostToDevice));
 
-	checkCudaErrors(cudaMemcpy(vertexA, geom.vertexA, arraySize * sizeof(double3), cudaMemcpyHostToDevice));
+	checkCudaErrors(cudaMemcpy(vertexA, geom.vertexA, arraySize * sizeof(float3), cudaMemcpyHostToDevice));
 
-	checkCudaErrors(cudaMemcpy(edgeBA, geom.edgeBA, arraySize * sizeof(double3), cudaMemcpyHostToDevice));
+	checkCudaErrors(cudaMemcpy(edgeBA, geom.edgeBA, arraySize * sizeof(float3), cudaMemcpyHostToDevice));
 
-	checkCudaErrors(cudaMemcpy(edgeCA, geom.edgeCA, arraySize * sizeof(double3), cudaMemcpyHostToDevice));
+	checkCudaErrors(cudaMemcpy(edgeCA, geom.edgeCA, arraySize * sizeof(float3), cudaMemcpyHostToDevice));
 
-	checkCudaErrors(cudaMemcpy(center, geom.center, arraySize * sizeof(double3), cudaMemcpyHostToDevice));
+	checkCudaErrors(cudaMemcpy(center, geom.center, arraySize * sizeof(float3), cudaMemcpyHostToDevice));
 
 	checkCudaErrors(cudaMemcpy(area, geom.area, arraySize * sizeof(double), cudaMemcpyHostToDevice));
 }
